refactor(google_vertex): extract request/response transforms to google_vertex_transforms.py
Phase 0c-2 thinning (slice 2): move the 7 pure transform functions (model-id/
location resolution, OpenAI<->Vertex tool/message translation) into
google_vertex_transforms.py — zero intra-file calls, zero shared state, no
external importers. client 1411 -> 1101 (1785 -> 1101, -38% across 0c-2).
Verified against the full local suite (zero regressions).
